+++ src/vector/float.rs
@@ -452,7 +452,33 @@ where
         self - other * self.dot(other)
     }
 
-    /// Returns the reflection of `self` through `normal`.
+    /// Returns the vector reflection of `self` off the surface with normal
+    /// `normal`.
+    ///
+    /// This assumes `normal` can be normalized.
+    ///
+    /// # Panics
+    ///
+    /// When debug assertions are enabled:
+    ///
+    /// Panics if `normal` cannot be normalized.
+    #[inline]
+    #[must_use]
+    #[track_caller]
+    pub fn reflect_off(self, normal: Self) -> Self {
+        let normal_length_squared = normal.length_squared();
+
+        debug_assert!(
+            normal_length_squared.recip().is_finite(),
+            "`normal` cannot be normalized: {self:?}.reflect_off({normal:?})"
+        );
+
+        let dot = self.dot(normal);
+        self - normal * ((dot + dot) / normal_length_squared)
+    }
+
+    /// Returns the vector reflection of `self` off the surface with normal
+    /// `normal`.
     ///
     /// This assumes `normal` is normalized.
     ///
@@ -464,13 +490,14 @@ where
     #[inline]
     #[must_use]
     #[track_caller]
-    pub fn reflect(self, normal: Self) -> Self {
+    pub fn reflect_off_normalized(self, normal: Self) -> Self {
         debug_assert!(
             normal.is_normalized(),
-            "normal is not normalized: {self:?}.reflect({normal:?})"
+            "`normal` is not normalized: {self:?}.reflect_off_normalized({normal:?})"
         );
 
-        self - normal * (T::as_from(2.0) * self.dot(normal))
+        let dot = self.dot(normal);
+        self - normal * (dot + dot)
     }
 
     /// Returns the vector refraction of `self` through `normal` and `eta`.
@@ -1492,6 +1519,34 @@ where
     #[must_use]
     pub fn sin_cos(self) -> (Self, Self) {
         specialize!(<T as FloatVectorBackend<N, A>>::sin_cos(self))
+    }
+
+    /// Returns the vector reflection of `self` off the surface with normal
+    /// `normal`.
+    ///
+    /// This assumes `normal` is normalized.
+    ///
+    /// This function has been deprecated and replaced by
+    /// [`reflect_off_normalized`].
+    ///
+    /// # Panics
+    ///
+    /// When debug assertions are enabled:
+    ///
+    /// Panics if `normal` is not normalized.
+    ///
+    /// [`reflect_off_normalized`]: Self::reflect_off_normalized
+    #[inline]
+    #[must_use]
+    #[track_caller]
+    #[deprecated(since = "0.18.2", note = "replaced by `reflect_off_normalized`")]
+    pub fn reflect(self, normal: Self) -> Self {
+        debug_assert!(
+            normal.is_normalized(),
+            "normal is not normalized: {self:?}.reflect({normal:?})"
+        );
+
+        self - normal * (T::as_from(2.0) * self.dot(normal))
     }
 }
 
@@ -3099,6 +3154,74 @@ mod tests {
     }
 
     #[test]
+    fn test_reflect_off() {
+        for_types!(|T: PrimitiveFloat, A| {
+            assert_test_eq!(
+                Vector::<2, T, A>::new(3.0, 2.0).reflect_off(Vector::ONE),
+                Vector::<2, T, A>::new(-2.0, -3.0),
+                abs <= 1e-5
+            );
+            assert_test_eq!(
+                Vector::<3, T, A>::new(3.0, 2.0, 4.0).reflect_off(Vector::ONE),
+                Vector::<3, T, A>::new(-3.0, -4.0, -2.0),
+                abs <= 1e-5
+            );
+            assert_test_eq!(
+                Vector::<4, T, A>::new(3.0, 2.0, 4.0, 5.0).reflect_off(Vector::ONE),
+                Vector::<4, T, A>::new(-4.0, -5.0, -3.0, -2.0),
+                abs <= 1e-5
+            );
+
+            assert_debug_panic!(
+                Vector::<2, T, A>::new(3.0, 2.0).reflect_off(Vector::<2, T, A>::ZERO)
+            );
+            assert_debug_panic!(
+                Vector::<3, T, A>::new(3.0, 2.0, 4.0).reflect_off(Vector::<3, T, A>::ZERO)
+            );
+            assert_debug_panic!(
+                Vector::<4, T, A>::new(3.0, 2.0, 4.0, 5.0).reflect_off(Vector::<4, T, A>::ZERO)
+            );
+        });
+    }
+
+    #[test]
+    fn test_reflect_off_normalized() {
+        for_types!(|T: PrimitiveFloat, A| {
+            assert_test_eq!(
+                Vector::<2, T, A>::new(3.0, 2.0)
+                    .reflect_off_normalized(Vector::<2, T, A>::ONE.normalize()),
+                Vector::<2, T, A>::new(-2.0, -3.0),
+                abs <= 1e-5
+            );
+            assert_test_eq!(
+                Vector::<3, T, A>::new(3.0, 2.0, 4.0)
+                    .reflect_off_normalized(Vector::<3, T, A>::ONE.normalize()),
+                Vector::<3, T, A>::new(-3.0, -4.0, -2.0),
+                abs <= 1e-5
+            );
+            assert_test_eq!(
+                Vector::<4, T, A>::new(3.0, 2.0, 4.0, 5.0)
+                    .reflect_off_normalized(Vector::<4, T, A>::ONE.normalize()),
+                Vector::<4, T, A>::new(-4.0, -5.0, -3.0, -2.0),
+                abs <= 1e-5
+            );
+
+            assert_debug_panic!(
+                Vector::<2, T, A>::new(3.0, 2.0).reflect_off_normalized(Vector::<2, T, A>::ONE)
+            );
+            assert_debug_panic!(
+                Vector::<3, T, A>::new(3.0, 2.0, 4.0)
+                    .reflect_off_normalized(Vector::<3, T, A>::ONE)
+            );
+            assert_debug_panic!(
+                Vector::<4, T, A>::new(3.0, 2.0, 4.0, 5.0)
+                    .reflect_off_normalized(Vector::<4, T, A>::ONE)
+            );
+        });
+    }
+
+    #[test]
+    #[expect(deprecated)]
     fn test_reflect() {
         for_types!(|T: PrimitiveFloat, A| {
             assert_test_eq!(