+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- [#102](https://github.com/Noam2Stein/ggmath/pull/102): Replace
+  `Vector::reflect` with `reflect_off` and `reflect_off_normalized`
+
 ## [0.18.1] - 2026-09-16
 
 ### Added